+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Removed:
 
 Improved/Fixed:
 
+- fixed inlining including libs not working, now it looks for the file relative to the file not the project root
 - Queue and stacks vars are now automatically destroyed when they go out of scope
 - Enforced move semantics, now they apply
 - fixed queue stack param passing